refactor: PermutationMatrixのビルダの実装を分離する
